BUGFIX: handle digest_unlink to a non-cached inode
If kernfs is restarted. it is possible that an inode doesn't have a
in-memory copy when handling its unlink digest request.
Current workaround is to create a new in-memory copy, which will be
reset to invalid state, marked dirty and flushed (invalid state) back to NVM when
the current digest completes.