Fix fragmentation simulation.
- First, we statically select the number of blocks to allocate. This is
the mininum of the number requested blocks and the max average
continuity of the layout score.
- Then, we randomly select a free region in the file system. This is
done by performing a random tree descent (as the free list is managed as
a red-black tree in memory).
- Then, if the free region is >> the number of requested blocks, we
split the region in half and allocate our requested blocks in the middle
of the region.

NOTE: this does require large allocations at a time from LibFS to be
effective (ie large fallocates/writes), otherwise the layout score will
just end up being zero. Should be fine for our tests, however, since we
ensure a minimum allocation size