Add the script to draw radar graph.
Note that you have to modify the python script itself (a global boolean
flag `DoGlobal`) to generate two different graphs (one for global data
structure and another one for per-file data structure).
The script will show a preview of the radar chart then save the chart
together with legends to a pdf file.